requirements_test.txt
coverage
coveralls
django-nose
mock
nose-exclude
